apps/api/pyproject.toml
[project]
name = "hephae-api"
version = "0.1.0"
description = "Unified Hephae API — serves both web and admin frontends"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "google-adk>=1.0.0",
    "google-generativeai>=0.8.0",
    "firebase-admin>=6.6.0",
    "google-cloud-bigquery>=3.27.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-tasks>=2.16.0",
    "playwright>=1.48.0",
    "httpx>=0.28.0",
    "pdfplumber>=0.11.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0",
    "resend>=2.5.0",
    "python-dotenv>=1.0.0",
    "hephae-common",
    "hephae-db",
    "hephae-integrations",
    "hephae-agents",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hephae_api"]

[tool.ruff]
target-version = "py312"
line-length = 120
